move functions relative to delaunay method to specific module, remove all code related to convolution method
